api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.35.0
redis==5.0.8
certifi>=2024.2.2
orjson==3.11.1
pydantic==2.11.9
discord.py==2.3.2
stripe==7.8.0